// client/src/app/interfaces/employeeEducation.ts
export interface IEmployeeEducation {
  employeeID: number;
  educationLevel: string;
  educationType: string;
  educationCenterName: string;
  educationProfile: string;
  educationDegree: string;
  educationDateEnd: string;
  educationDiplomaNumber: string;
  educationDiplomaDate: string;
  _id: string;
  updatedAt?: string;
  createdAt?: string;
  selected?: boolean;
  editable?: boolean;
}